Fold only parent class into only child class
